# baggage: Member.String encodes only when necessary