--- a/baggage/baggage_test.go
+++ b/baggage/baggage_test.go
@@ -501,13 +501,7 @@ func TestBaggageString(t *testing.T) {
 			// Meaning, US-ASCII characters excluding CTLs, whitespace,
 			// DQUOTE, comma, semicolon, and backslash. All excluded
 			// characters need to be percent encoded.
-			//
-			// Ideally, the want result is:
-			// out: "foo=%00%01%02%03%04%05%06%07%08%09%0A%0B%0C%0D%0E%0F%10%11%12%13%14%15%16%17%18%19%1A%1B%1C%1D%1E%1F%20!%22#$%25&'()*+%2C-./0123456789:%3B<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[%5C]^_%60abcdefghijklmnopqrstuvwxyz{|}~%7F",
-			// However, the following characters are escaped:
-			// !#'()*/<>?[]^{|}
-			// It is not necessary, but still provides a correct result.
-			out: "foo=%00%01%02%03%04%05%06%07%08%09%0A%0B%0C%0D%0E%0F%10%11%12%13%14%15%16%17%18%19%1A%1B%1C%1D%1E%1F%20%21%22%23$%25&%27%28%29%2A+%2C-.%2F0123456789:%3B%3C=%3E%3F@ABCDEFGHIJKLMNOPQRSTUVWXYZ%5B%5C%5D%5E_%60abcdefghijklmnopqrstuvwxyz%7B%7C%7D~%7F",
+			out: "foo=%00%01%02%03%04%05%06%07%08%09%0A%0B%0C%0D%0E%0F%10%11%12%13%14%15%16%17%18%19%1A%1B%1C%1D%1E%1F%20!%22#$%25&'()*+%2C-./0123456789:%3B<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[%5C]^_`abcdefghijklmnopqrstuvwxyz{|}~%7F",
 			baggage: baggage.List{
 				"foo": {Value: func() string {
 					// All US-ASCII characters.