@dmathieu 2023-12-21
Should you run benchstat to compare the benchmarks between main and your branch?