@pellared 2023-12-21
> Should you run benchstat to compare the benchmarks between main and your branch?

I can do that today in a few hours or tomorrow morning. 
The result should be similar (maybe a little faster as there should be a few CPU cycles less). I intentionally almost copied the code from `url.PathEscape`. I wanted to add the benchmarks if someone would want to improve it in future.